+++ package.json
@@ -10,16 +10,14 @@
     "dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "test": "jest",
-    "prisma:generate": "prisma generate",
-    "prisma:migrate": "prisma migrate dev",
-    "prisma:studio": "prisma studio",
-    "postinstall": "prisma generate"
+    "typeorm": "typeorm-ts-node-commonjs -d ./src/data-source.ts",
+    "migration:generate": "npm run typeorm -- migration:generate ./src/migrations/ -d ./src/data-source.ts",
+    "migration:run": "npm run typeorm -- migration:run",
+    "migration:revert": "npm run typeorm -- migration:revert"
   },
   "keywords": [
     "mezon",
-    "bot",
-    "card-game",
-    "21"
+    "bot"
   ],
   "author": "",
   "license": "ISC",
@@ -32,13 +30,14 @@
     "@nestjs/platform-express": "^11.1.1",
     "@nestjs/serve-static": "^5.0.3",
     "@nestjs/testing": "^11.1.1",
-    "@prisma/client": "6.5.0",
+    "@nestjs/typeorm": "^11.0.0",
     "@types/dotenv": "^6.1.1",
     "@types/node": "^20.11.19",
     "dotenv": "^16.5.0",
     "ioredis": "^5.6.1",
     "kysely": "^0.28.2",
     "mezon-sdk": "2.7.96",
+    "mysql2": "^3.14.5",
     "prisma": "^6.7.0",
     "prisma-kysely": "^1.8.0",
     "reflect-metadata": "^0.2.2",
@@ -50,6 +49,7 @@
     "@nestjs/cli": "^11.0.7",
     "@types/jest": "^29.5.12",
     "jest": "^29.7.0",
-    "ts-jest": "^29.1.2"
+    "ts-jest": "^29.1.2",
+    "typeorm": "^0.3.26"
   }
 }
