+++ src/bot/commands/index.ts
