+++ src/common/enums/bot.enum.ts
@@ -1,1 +1,18 @@
 export const COMMAND_PREFIX = ['*'];
+
+export enum GameStatus {
+   WAITING = 'WAITING',
+   BETTING = 'BETTING',
+   ROLLING = 'ROLLING',
+   FINISHED = 'FINISHED',
+   CANCELLED = 'CANCELLED',
+}
+
+export enum DiceSymbol {
+   FISH = 'FISH',
+   SHRIMP = 'SHRIMP',
+   CRAB = 'CRAB',
+   GOURD = 'GOURD',
+   ROOSTER = 'ROOSTER',
+   DEER = 'DEER',
+}
